Saját optimalizáló eljárás mergelése main-be (nem végleges, mostani legjobb állapot)

+++ backend/cpp/main_plus_viterbi_fingering_optimization/optimization.h
@@ -11,6 +11,6 @@ public:
 
 private:
     std::vector<InputNotes> notes;
-    double CalculateCenter(const int currentIdx) const;
+    double CalculateCenter(const size_t currentIdx) const;
     double ExtraCost(const double currentCenter, const NotePosition &nextPos, const NotePosition &prevPos) const;
 };
